requirements.txt
click==8.0.2
Flask~=2.0.1
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
Werkzeug~=2.0.1

bcrypt~=3.2.0
Pillow~=8.1.0
psycopg2~=2.8.6